refactor: Implement poetry
