--- todos.txt
@@ -1,3 +1,0 @@
-learn
-earn
-play
